+++ modules/bibcite_bibtex/tests/src/Kernel/BibtexDecodeTest.php
@@ -16,7 +16,7 @@ class BibtexDecodeTest extends FormatDecoderTestBase {
    *
    * @var array
    */
-  public static $modules = [
+  protected static $modules = [
     'system',
     'user',
     'serialization',