+++ modules/bibcite_endnote/tests/src/Kernel/TaggedDecodeTest.php
@@ -16,7 +16,7 @@ class TaggedDecodeTest extends FormatDecoderTestBase {
    *
    * @var array
    */
-  public static $modules = [
+  protected static $modules = [
     'system',
     'user',
     'serialization',