+++ modules/bibcite_export/tests/src/Functional/EntityExportTest.php
@@ -15,7 +15,7 @@ class EntityExportTest extends BrowserTestBase {
   /**
    * {@inheritdoc}
    */
-  public static $modules = [
+  protected static $modules = [
     'bibcite_export_test',
   ];