+++ modules/bibcite_import/tests/src/Kernel/ImportBasicTest.php
@@ -19,7 +19,7 @@ class ImportBasicTest extends KernelTestBase {
    *
    * @var array
    */
-  public static $modules = [
+  protected static $modules = [
     'system',
     'user',
     'text',