+++ modules/bibcite_marc/tests/src/Kernel/MarcDecodeTest.php
@@ -16,7 +16,7 @@ class MarcDecodeTest extends FormatDecoderTestBase {
    *
    * @var array
    */
-  public static $modules = [
+  protected static $modules = [
     'system',
     'user',
     'serialization',