+++ modules/bibcite_ris/tests/src/Kernel/RisDecodeTest.php
@@ -16,7 +16,7 @@ class RisDecodeTest extends FormatDecoderTestBase {
    *
    * @var array
    */
-  public static $modules = [
+  protected static $modules = [
     'system',
     'user',
     'serialization',